requirements.txt
gdown>=4.7.1
openpyxl==3.1.2
sqlalchemy>=2.0.35
schedule==1.2.0
python-dotenv==1.0.0
requests==2.32.5
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiogram>=3.0.0

